requirements.txt
build123d==0.10.0
ocp-tessellate==3.2.2
python-solvespace==3.0.8
PyQt6==6.11.0
PyOpenGL==3.1.10
numpy==2.4.4
